requirements.txt
numpy
scipy
tensorflow>=2.1
absl-py
